+++ src/server/utils/api.ts
@@ -48,7 +48,7 @@ export default class API {
   public static readonly v460 = API.fromSimpleString('4.6.0');
   public static readonly v480 = API.fromSimpleString('4.8.0');
   public static readonly v490 = API.fromSimpleString('4.9.0');
-
+  public static readonly v510 = API.fromSimpleString('5.1.0');
 
   public static fromVersionString(versionString: string): API {
     let version = semver.valid(versionString)